% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{apply.flow.filters}
\alias{apply.flow.filters}
\title{apply.flow.filters}
\usage{
apply.flow.filters(
  gh,
  frame.sf = NULL,
  directed = F,
  min.tie.str = NULL,
  tie.str.deciles = 5,
  min.flows = 10,
  max.dst = units::set_units(10, "miles")
)
}
\arguments{
\item{gh}{a graph, as setup by \code{sfg2gh}}

\item{directed}{whether to leave directed or make undirected (which aggregates
trips to/from the same tracts, rather than duplicating the O-D pair for each
direction.)}

\item{min.str}{If null, filters to top quartile after other filters applied}
}
\description{
Apply flow filters to a graph, as generated from \code{sfg2gh} and spatialized with
\code{spatialize.graph}. Crops to bounds.sf, filters by distance, total trips, and flow
strength based on parameters. Note that cropping to bounds sf is done to retain
edges that start outside of map area but ends within. This is done by cropping by
edges, which also creates a number of unnamed/NA-named nodes, which are artifacts
of nodes outside of cropping area when edges are partially inside.
}
