% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{attenuated.power.law}
\alias{attenuated.power.law}
\title{attenuated power law}
\usage{
attenuated.power.law(d, pb = 1, gamma = 1, alpha = 1)
}
\description{
"Unlike the simple power law model....the attenuated variant exhibits
negative local curvature. This model thus tends to produce both a high proportion
of long-range edges and a high degree of local cohesion (the scale of these
effects being dependent upon α). Where (αd)γ » 1, both variants exhibit similar
behavior."
}
