% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{build.pairwise.dst.matrix}
\alias{build.pairwise.dst.matrix}
\title{build.pairwise.dst.matrix}
\usage{
build.pairwise.dst.matrix(sfx, decay = F, h0 = 32.2, ...)
}
\arguments{
\item{sfx}{sf object to generate distances from}

\item{decay}{Whether to get decayed distances with bisq weights}

\item{h0}{if \code{decay} is \code{TRUE}, the distance in KM at which spatial weights are 0.}
}
\value{
distance or spatial weights matrix. If distance, matrix is of numerics
representing KMs
}
\description{
Builds a pairwise distance matrix from an sf object. Option to get bisq distance
wights. Mostly just wraps \code{sf::st_distance}. See
http://mjh4.blogspot.com/2012/09/ergm-edgecov-and-dyadcov-specifications.html or
https://cran.r-project.org/web/packages/ergm/vignettes/ergm-term-crossRef.html for
\code{edgecov}/\code{dyadcov} documentation.
}
