% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-della-fcns.R
\name{Della.wrapper_cz.ergm.variations}
\alias{Della.wrapper_cz.ergm.variations}
\title{Della.wrapper_cz.ergm.variations}
\usage{
Della.wrapper_cz.ergm.variations(
  cz,
  directed = F,
  flow.floor = 1,
  drop.loops = T,
  pop.floor = 500,
  dropbox.dir = "/scratch/gpfs/km31/",
  ddir = "/home/km31/all/divseg/.input-datasets/",
  seed = 1234
)
}
\description{
A metawrapper. Preps graphs, runs a series of ERGM models, saves
visualizations, and saves all three sets of objects. A lot is hardcoded for
now. In particular, NHPN hwy data should be sent along with the fcn using the
\code{rslurm::slurm_apply} \code{add_objects} argument (called \code{nhpn}), as well as a
list of ergm specifications titled \code{model.variations.}
}
\details{
Will maybe change so it requires a proximity matrix or list of prx.matrices
at some point..
}
