% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spW-fcns.R
\name{Della.wrapper_flow.weights_by.rt}
\alias{Della.wrapper_flow.weights_by.rt}
\title{Della.wrapper_flow.weights_by.rt}
\usage{
Della.wrapper_flow.weights_by.rt(
  cz_id = NULL,
  cbsa_id = NULL,
  agg2tracts = T,
  drop.loops = T,
  weight.floor = 0.01,
  sfg.dir,
  year = "2019",
  save.dir = NULL,
  list.colms = T
)
}
\arguments{
\item{drop.loops}{Whether to drop loops (where origin==destination)}

\item{weight.floor}{flow weight floor. Minimum percent of incoming/visited
trips between tracts to be included. 0.1\% by default.}

\item{list.colms}{whether to return list columns (or tibble colms)}
}
\description{
Wrapper function to generate all flow weights by either CZ or CBSA.
}
