% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spW-fcns.R
\name{Della.wrapper_flow.weights_dst.cutoff}
\alias{Della.wrapper_flow.weights_dst.cutoff}
\title{Della.wrapper_flow.weights_dst.cutoff}
\usage{
Della.wrapper_flow.weights_dst.cutoff(
  cz,
  agg2tracts = T,
  drop.loops = T,
  weight.floor = 0.001,
  sfg.dir,
  year = "2019",
  save.dir = NULL
)
}
\arguments{
\item{drop.loops}{Whether to drop loops (where origin==destination)}

\item{weight.floor}{flow weight floor. Minimum percent of incoming/visited
trips between tracts to be included. 0.1\% by default.}
}
\description{
Wrapper function to generate all flow weights. In addition to parameters,
keep a spatial weights object \code{spws} in global env. This is a a proximity
list as calculated with the \verb{creating SpWs} scripts and needs the
\code{below.cutoff} list-column with tract/blockgroup geoids within the distance
maximum.
}
