% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-della-fcns.R
\name{Della.wrapper_sif.variations}
\alias{Della.wrapper_sif.variations}
\title{Della.wrapper_sif.variations}
\usage{
Della.wrapper_sif.variations(cz, seed = 1234)
}
\description{
Not a graceful fcn; include \code{model.variations} & \code{prx.mats} in
\code{include_objects} rslurm call.
}
