% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spW-fcns.R
\name{dsts.below.threshold}
\alias{dsts.below.threshold}
\title{most fcns that would be here are just moved to the sending Della job scripts
that generate the weights or composites.}
\usage{
dsts.below.threshold(sfx, i, j.colm = "nbs")
}
\arguments{
\item{sfx}{an sf object to selectively get distances from i, probably
containing point geometries.}

\item{i}{row # index of given tract/centroid in \code{sfx}}

\item{j.colm}{list-column of geoids in \code{sfx} to get distance from}
}
\description{
In R/creating SpWs/sending della jobs/
dsts.below.threshold
}
\details{
Quick fcn with hardcoded items, including assumption of meters from crs length
unit
}
\examples{
\dontrun{
ctrs <- cts \%>\% st_centroid()
dst.nbs <-
 st_is_within_distance(ctrs
                         , dist =
                         units::set_units(20,'miles'))
cts$nbs <- map(dst.nbs, ~tmp[., ]$geoid)
ctrs \%>\%
mutate(dists =
         map(1:nrow(ctrs)
          ,~dsts.below.threshold(ctrs, ., 'nbs')))
}

}
