% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{ergm.clean.n.convert}
\alias{ergm.clean.n.convert}
\title{ergm.clean.n.convert}
\usage{
ergm.clean.n.convert(gh, pop.floor = 0)
}
\arguments{
\item{population.floor}{Applies a population floor. Will assume a node-based \code{pop}
column to filter by}
}
\value{
a \code{network} object for use with statnet libraries.
}
\description{
the ergms packages can't handle some types of information, like factors and NAs.
This applies final cleans neccessary for an ERGM to run without error and then
coerces to \code{network} object.
}
