% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{exponential.decay}
\alias{exponential.decay}
\title{inverse exponential decay}
\usage{
exponential.decay(d, pb = 1, alpha = 0.5, ...)
}
\arguments{
\item{d}{distance}
}
\description{
One of the functional forms posited for the SIF in Daraganova 2012 and Butts and
Acton 2011.
}
\details{
Fd(x = pb exp(–αx), α >=0 scaling param
}
