% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-della-fcns.R
\name{fn2tags}
\alias{fn2tags}
\title{fn2tags}
\usage{
fn2tags(fn, tag = c("region", "formula", "all"))
}
\arguments{
\item{fn}{a filename of saved model .rds, and it's directory.}
}
\description{
Gets region and/or formula tags from a filename saved accoring to conventions
used by Della wrapper fcns here.
}
