% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spW-fcns.R
\name{get.dist.weighted.composite}
\alias{get.dist.weighted.composite}
\title{get.dist.weighted.composite}
\usage{
get.dist.weighted.composite(
  i,
  x,
  value.col = "value",
  weight.col = "weight",
  dist.decay.fcn,
  dist.col = "dists",
  spatial.weights = spws,
  ...
)
}
\arguments{
\item{i}{a given tract geoid}

\item{x}{full ct data to subset from}

\item{value.col}{Attribute column name in \code{x} to calculate weighted avg of
adjacent tracts for.}

\item{weight.col}{column name in \code{x} for non spatial weights. Probably population
or households}

\item{dist.decay.fcn}{a distance decay or proximity fcn to convert distances to
spatial weights}

\item{dist.col}{string for column name in \code{spws} that pairs geoids with distances
from i}

\item{...}{passed on to \code{dst.decay.fcn}}

\item{spws}{spatial weight matrices with a geoid column and other list columns for
spatial weights}
}
\description{
Given an area geoid, a set of attribute variables for all areas, and a list-column
of spatial weights between areas, calculate a "spatial composite" that
incorporates neighbors' attributes, spatial weights, and population or other
weights.
}
