% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-viz-fcns.R
\name{get.div.layers}
\alias{get.div.layers}
\title{get.div.layers}
\usage{
get.div.layers(
  bounds.sf,
  map.buffer = units::set_units(5, "miles"),
  dropbox.dir = Sys.getenv("drop_dir"),
  subdir = "shapefiles/nhpn/"
)
}
\description{
Wraps visaux helpers with very specific settings for combining everything.
}
