% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spW-fcns.R
\name{get.flow.weights.within.distance}
\alias{get.flow.weights.within.distance}
\title{get.flow.weights.within.distance}
\usage{
get.flow.weights.within.distance(
  i,
  flow.type = c("incoming", "visited"),
  drop.loops = T,
  prox.col = "below.cutoff",
  flow.counts = sfg,
  spatial.weights = spws,
  weight.floor = 0.001
)
}
\arguments{
\item{i}{a given tract geoid}

\item{flow.type}{Whether to get weights based on all visitors to i from other
proximate areas (default); or based on all proximiate areas visited by people in
area i.}

\item{drop.loops}{Whether to drop loops (where origin==destination)}

\item{prox.col}{name for list-column in \code{spws} with all areas within distance
cutoff from each other}

\item{flow.counts}{data.frame with origin/destination/estimated visits (n).}

\item{weight.floor}{flow weight floor. Minimum percent of incoming/visited trips
between tracts to be included. 0.1\% by default.}
}
\description{
Get "flow weights," i.e., proportion of visitor/visited from other tracts within
distance threshold. Relies on precalculated distance matrix. Rounds to 4 digits
for sparseness.
}
