% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{get.normalized.undirected.connectedness}
\alias{get.normalized.undirected.connectedness}
\title{get.normalized.undirected.connectedness}
\usage{
get.normalized.undirected.connectedness(x, y, el, flow.colm = "n")
}
\arguments{
\item{x, y}{node ids}

\item{el}{edge list}

\item{flow.colm}{column containing edge weights/flows between nodes.}
}
\description{
Gets a measure of "tie strength" of total visits between tracts, scaled by total
flows into and out of each tract. Takes individual node/tract identifiers and an
edgelist containing all flows in between them.
}
