% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{is.in.city.center}
\alias{is.in.city.center}
\title{is.in.city.center}
\usage{
is.in.city.center(sfx, cz = NULL, filter2cc = FALSE, ...)
}
\arguments{
\item{sfx}{a tidy graph with node centroids in a geometry column}

\item{cz}{a cz corresponding with the graph; if left \code{NULL}, fcn will assume
a column with the cz id exists for nodes already}

\item{filter2cc}{Whether or not to filter or just add column}
}
\value{
a tidy graph with new \code{in.cc} column for nodes, based on whether or
not they are in a population center. If filter2cc is \code{TRUE}, only keeps
nodes within population center
}
\description{
Given a tidy graph with tract centroid geometries, get each nodes' status as
within or outside the city center, defined as the largest Place by population
in the commuting zone. References a dataset bundled with \code{divM} that already
pairs these population centers with commuting zones.
}
