% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{load.sfg}
\alias{load.sfg}
\title{load.sfg}
\usage{
load.sfg(
  .czs,
  min.flows = 10,
  tracts.or.groups = c("ct", "bg"),
  trim.loops = T,
  year = "2019",
  ddir = Sys.getenv("drop_dir"),
  sfg.dir = "sfg-processed/orig_dest_annual/",
  ...
)
}
\arguments{
\item{min.flows}{flow floor to immediately filter to. Default 10.}

\item{tracts.or.groups}{Aggregate to census tract (ct) or leave at block group (bg)}

\item{trim.loops}{whether to trim loops, where origin==dest. Default true.}

\item{year}{year to load for. Default 2019.}

\item{ddir, sfg.dir}{location of safegraph data, saved by cz.}
}
\description{
load.sfg
}
