% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-della-fcns.R
\name{model2tidy}
\alias{model2tidy}
\title{model2tidy}
\usage{
model2tidy(
  fn,
  mdir = "/scratch/gpfs/km31/ergms/models/",
  svpath = "/scratch/gpfs/km31/ergms/model summaries/cz-ergms.csv"
)
}
\arguments{
\item{fn, mdir}{a filename of saved model .rds, and it's directory.}

\item{svpath}{Path to a .csv to which to write mass model summary}
}
\description{
Loads a model .rds saved with the della script, extracts some diagnostics and
coefficient estimates, and saves to a running .csv with all model outputs.
Designed to map across filenames
}
