% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{neg.exp}
\alias{neg.exp}
\title{neg.exp}
\usage{
neg.exp(d, ...)
}
\arguments{
\item{d}{distance}
}
\description{
e^-distance. Used by Massey and Denton for the Clustering dimension of
segregation. Notably they include tract where i==j and use an estimated
self-distance
}
