% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{power.law.loss.fcn}
\alias{power.law.loss.fcn}
\title{for the parameters for the spatial interaction fucntion
power.law.loss.fcn}
\usage{
power.law.loss.fcn(par)
}
\arguments{
\item{par}{vector of parameters passed onto \code{power.law}}
}
\description{
Loss function that wraps the \code{power.law} fcn. Uses the \code{dst.freq} object, as
returned from \code{relative.tie_dst.frequency}. This function is basically copied
from https://rpubs.com/YaRrr/MLTutorial
}
