% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{power.law}
\alias{power.law}
\title{power law}
\usage{
power.law(d, pb = 1, gamma = 1, alpha = 1, ...)
}
\arguments{
\item{d}{distance}
}
\description{
One of the functional forms posited for the SIF in Daraganova 2012 and Butts and
Acton 2011.
}
