% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spW-fcns.R
\name{region2identifiers}
\alias{region2identifiers}
\title{region2identifiers}
\usage{
region2identifiers(region.type, region.id)
}
\description{
Turns region.type and region.id into a cleaned & concatenated set of
identifiers, which I use for partial saves.
}
