% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{relative.tie_dst.frequency}
\alias{relative.tie_dst.frequency}
\title{relative.tie_dst.frequency}
\usage{
relative.tie_dst.frequency(
  sfn = NULL,
  sfe = NULL,
  gh = NULL,
  n.bins = 50,
  dst.mat = NULL
)
}
\arguments{
\item{sfn, sfe, gh}{Spatial nodes, spatial edges or graph object. Save load time by
providing spatial nodes / edges; otherwise just supply graph gh.}

\item{n.bins}{number of distance bins}
}
\description{
Gets distribution of ties relative to distribution of centroids. Because centroids
are not an even continuous distribution, this might show a more meaningful spatial
distribution of ties.
}
\details{
Implements technique in Daraganova 2012, referred to as  "binned relative
frequency of ties against distance interval"
}
