% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{rid2sfg}
\alias{rid2sfg}
\title{rid2sfg}
\usage{
rid2sfg(cz = NULL, cbsa = NULL, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=load.sfg]{load.sfg}}
  \describe{
    \item{\code{min.flows}}{flow floor to immediately filter to. Default 10.}
    \item{\code{trim.loops}}{whether to trim loops, where origin==dest. Default true.}
    \item{\code{year}}{year to load for. Default 2019.}
    \item{\code{ddir}}{location of safegraph data, saved by cz.}
    \item{\code{sfg.dir}}{location of safegraph data, saved by cz.}
    \item{\code{tracts.or.groups}}{Aggregate to census tract (ct) or leave at block group (bg)}
  }}
}
\description{
rid2sfg
}
