% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{setup.gh.wrapper}
\alias{setup.gh.wrapper}
\title{setup.gh.wrapper}
\usage{
setup.gh.wrapper(sfx = NULL, cz = NULL, cbsa = NULL, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=sfx2sfg]{sfx2sfg}}, \code{\link[=sfg2gh]{sfg2gh}}, \code{\link[=spatialize.graph]{spatialize.graph}}, \code{\link[=apply.flow.filters]{apply.flow.filters}}
  \describe{
    \item{\code{sfx}}{an sf object within bounds of which to load all flows.}
    \item{\code{sfg}}{OD dataframe}
    \item{\code{directed}}{whether to leave directed or make undirected (which aggregates
trips to/from the same tracts, rather than duplicating the O-D pair for each
direction.)}
    \item{\code{gh}}{a graph, as setup by \code{sfg2gh}}
    \item{\code{frame.sf}}{an \code{sf} object to crop to and use CRS from. Nodes outside the bbox of
this area will be trimmed. Defaults to Lambert conformal conic projection if
left null. (Why do I have all this complexity instead of just using coord_sf
when mappi ng? Think I need to revise.. I think b/c i felt it allowed for more
controlled trimming but..)}
    \item{\code{tracts.or.groups}}{Aggregate to census tract (ct) or leave at block group (bg)}
    \item{\code{min.tie.str}}{If not null, floor for tie strength}
    \item{\code{min.flows}}{minimum flows}
    \item{\code{tie.str.deciles}}{If not null, number of deciles of tie str to keep after all
other trims}
  }}
}
\description{
From a cz/cbsa identifier or an \code{sf} object, sets up a graph as all nodes surrounding
the area. Built to create a graph within a bounding box.
}
