% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{sfx2sfg}
\alias{sfx2sfg}
\title{sfx2sfg}
\usage{
sfx2sfg(sfx, year = 2019, ...)
}
\arguments{
\item{sfx}{an sf object within bounds of which to load all flows.}

\item{...}{
  Arguments passed on to \code{\link[=load.sfg]{load.sfg}}
  \describe{
    \item{\code{min.flows}}{flow floor to immediately filter to. Default 10.}
    \item{\code{trim.loops}}{whether to trim loops, where origin==dest. Default true.}
    \item{\code{year}}{year to load for. Default 2019.}
    \item{\code{ddir}}{location of safegraph data, saved by cz.}
    \item{\code{sfg.dir}}{location of safegraph data, saved by cz.}
    \item{\code{tracts.or.groups}}{Aggregate to census tract (ct) or leave at block group (bg)}
  }}
}
\description{
From a given \code{sf} object, get correspondence with commuting zones, and load all
safegraph data for trips with origins at those commuting zones. For now, just done
for annual, but we could make this monthly. If you want to start with just CZs,
just use sfg.seg::read.sfg.CZs.
}
