% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{spatialize.graph}
\alias{spatialize.graph}
\title{spatialize.graph}
\usage{
spatialize.graph(
  gh,
  frame.sf = NULL,
  tracts.or.groups = c("ct", "bg"),
  directed = F,
  year = 2019
)
}
\arguments{
\item{gh}{a graph, as setup by \code{sfg2gh}}

\item{tracts.or.groups}{Aggregate to census tract (ct) or leave at block group (bg)}

\item{directed}{whether to leave directed or make undirected (which aggregates
trips to/from the same tracts, rather than duplicating the O-D pair for each
direction.)}

\item{sfx}{an \code{sf} object to crop to and use CRS from. Nodes outside the bbox of
this area will be trimmed. Defaults to Lambert conformal conic projection if
left null. (Why do I have all this complexity instead of just using coord_sf
when mappi ng? Think I need to revise..)}
}
\description{
Spatializes a graph based on the node \code{name}s. Names should align with geoids for
either block groups or tracts (default).
}
