% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{apply.flow.filters}
\alias{apply.flow.filters}
\title{apply.flow.filters}
\usage{
apply.flow.filters(
  gh,
  tie.str.deciles = 5,
  directed = F,
  frame.sf = NULL,
  min.tie.str = NULL,
  min.flows = 10,
  max.dst = units::set_units(10, "miles")
)
}
\arguments{
\item{gh}{a graph, as setup by \code{sfg2gh}}

\item{tie.str.deciles}{If not null, number of deciles of tie str to keep after all
other trims}

\item{directed}{whether to leave directed or make undirected (which aggregates
trips to/from the same tracts, rather than duplicating the O-D pair for each
direction.)}

\item{min.tie.str}{If not null, floor for tie strenght}
}
\description{
Apply flow filters to a graph, as generated from \code{sfg2gh} and spatialized with
\code{spatialize.graph}. Crops to bounds.sf, filters by distance, total trips, and flow
strength based on parameters. Note that cropping to bounds sf is done to retain
edges that start outside of map area but ends within. This is done by cropping by
edges, which also creates a number of unnamed/NA-named nodes, which are artifacts
of nodes outside of cropping area when edges are partially inside.
}
