% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{bisq.dist2weights}
\alias{bisq.dist2weights}
\title{bisq.dist2weights}
\usage{
bisq.dist2weights(d, cutoff = dst.ceiling, ...)
}
\arguments{
\item{d}{distance}

\item{cutoff}{\code{H} in the formula, where weight is 0.}
}
\description{
Uses a Bisquare transformation to turn distances to spatial weights. Transfroms by
weight = (1-(ndist/H)^2)^2 for distances less than or equal to H, 0 otherwise.
Mimics Bi-square option in \code{lctools::moransI}.
}
