% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-viz-fcns.R
\name{flow.map.wrapper}
\alias{flow.map.wrapper}
\title{flow.map.wrapper}
\usage{
flow.map.wrapper(
  sfx,
  gh = NULL,
  edge.attr = "n",
  log.transform = T,
  map.buffer = units::set_units(5, "miles"),
  max.dst = units::set_units(5, "miles"),
  ...
)
}
\arguments{
\item{gh}{A graph object. Recreates based on \code{sfx} if not supplied}

\item{edge.attr}{Edge attribute to map by. 'n' by default. If generating graph
with these functions, tstr is tie strength for undirected graph and perc.to.dest
or perc.from.origin is for directed graph.}

\item{...}{
  Arguments passed on to \code{\link[=setup.gh.wrapper]{setup.gh.wrapper}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Big old wrapper function. Includes div layers: Place, US routes, interstates, and
water.
}
\details{
TODO: I think main issue here is that including outside of the bounds can mess up
what remains in the bounds... -> actually I think it's an issue of plot size and
the scale conversion to pixels.. Lines are thinner than a pixel and become
invisible depending on plot size. Need to think about how to handle that.
See: https://www.tidyverse.org/blog/2020/08/taking-control-of-plot-scaling/
}
