% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-viz-fcns.R
\name{get.div.layers}
\alias{get.div.layers}
\title{get.div.layers}
\usage{
get.div.layers(bounds.sf, ...)
}
\arguments{
\item{bounds.sf}{sf area that defines bbox to get division layers from.}

\item{...}{NHPN directory info passed onto \code{visaux::get.NHPN}}
}
\description{
Wraps visaux helpers with very specific settings for combining everything.
}
