% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIF-fcns.R
\name{neg.exp}
\alias{neg.exp}
\title{neg.exp}
\usage{
neg.exp(d, ...)
}
\arguments{
\item{d}{distance}
}
\description{
e^-distance.
}
\details{
Used by Massey and Denton to define a distance decay function used for various of
their possible indices for Clustering dimension of segregation. Notably they
include tract where i==j and use an non-zero within-tract distance in this case.
They estimate d_ii (average within tract distance for tract i) to be (.6a_i)^.5
"following White 1983." In contrast, this function here assumes no loops (i!=j),
and does not apply a within-tract weight.

I think this is a good, straightforward way to include decayed distance without
worrying about parameterizing.
}
