% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{setup.gh.wrapper}
\alias{setup.gh.wrapper}
\title{setup.gh.wrapper}
\usage{
setup.gh.wrapper(
  sfx,
  map.buffer = units::set_units(5, "miles"),
  max.dst = units::set_units(5, "miles"),
  min.flows = 10,
  min.str = NULL,
  directed = F,
  tracts.or.groups = c("ct", "bg"),
  year = 2019,
  base.dir = Sys.getenv("drop_dir"),
  sfg.dir = "sfg-processed/orig_dest_annual/",
  crs = "+proj=lcc +lon_0=-90 +lat_1=33 +lat_2=45"
)
}
\arguments{
\item{sfx}{an \code{sf} object to crop to and use CRS from. Nodes outside the bbox of
this area will be trimmed. Defaults to Lambert conformal conic projection if
left null. (Why do I have all this complexity instead of just using coord_sf
when mappi ng? Think I need to revise..)}

\item{min.str}{If null, filters to top quartile after other filters applied}

\item{directed}{whether to leave directed or make undirected (which aggregates
trips to/from the same tracts, rather than duplicating the O-D pair for each
direction.)}

\item{tracts.or.groups}{Aggregate to census tract (ct) or leave at block group (bg)}
}
\description{
From a \code{sf} object that will center graph, a buffer distance to define area around
center, and a set of trimming parameters, sets up a graph as all nodes surrounding
the area
}
