% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-and-ergm-fcns.R
\name{sfx2sfg}
\alias{sfx2sfg}
\title{sfx2sfg}
\usage{
sfx2sfg(
  eligible.sf,
  min.flows = 10,
  tracts.or.groups = c("ct", "bg"),
  trim.loops = T,
  year = 2019,
  base.dir = Sys.getenv("drop_dir"),
  sfg.dir = "sfg-processed/orig_dest_annual/"
)
}
\arguments{
\item{eligible.sf}{an sf object within bounds of which to load all flows.}
}
\description{
From a given \code{sf} object, get correspondence with commuting zones, and load all
safegraph data for trips with origins at those commuting zones. For now, just done
for annual, but we could make this monthly. If you want to start with just CZs,
just use sfg.seg::read.sfg.CZs.
}
